Move tests for OAuth to requester_test
